chore(test-pipeline): fix various authentication errors in functional tests
Services undergoing functional testing were unable to validate incoming JWT tokens because the issuer in the token did not match the expexted issuer. That is because the expected issuer was not being propagated to each service pod as an environment variable. This commit makes it so that the issuer, along with other important variables, are set in each pod's environment.
In addition, this commit introduces a wait for the webui pod to be ready before beginning functional tests.